changing types

+++ I2c_Imu/library.h
@@ -38,4 +38,4 @@
 #endif
 
 uint8_t innit_imu(); 
-void combine_data(uint8_t *data_array, uint16_t *clean_data);
+void combine_data(uint8_t *data_array, float *clean_data);
